tests(dichotomy): fix numpy cast error on python >= 3.10
fix this error:
```
numpy._core._exceptions._UFuncOutputCastingError: Cannot cast ufunc
'subtract' output from dtype('int64') to dtype('uint8') with casting
rule 'same_kind'
```

+++ tests/unit_tests/test_refinement/test_dichotomy.py
@@ -271,7 +271,7 @@ class TestRefinementMethod:
         # Remove Criteria.P2D_RIGHT_DISPARITY_OUTSIDE because it invalidates everything:
         cost_volumes["criteria"].data[
             Criteria.P2D_RIGHT_DISPARITY_OUTSIDE.is_in(cost_volumes["criteria"])
-        ] -= Criteria.P2D_RIGHT_DISPARITY_OUTSIDE
+        ] -= np.uint8(Criteria.P2D_RIGHT_DISPARITY_OUTSIDE)
         disp_map["validity"] = criteria.get_validity_dataset(cost_volumes["criteria"])["validity"]
         return disp_map
